Moved AudioFocusChangeListener to another file
Keep entities small... I'm trying xD
